requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cmake==3.21.0
commentjson==0.9.0
docker-pycreds==0.4.0
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.11
GitPython==3.1.40
Hydra==2.5
idna==3.6
imageio==2.16.1
importlib-metadata==6.7.0
lark-parser==0.7.8
lightning-utilities==0.10.0
multidict==6.0.4
numpy==1.21.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
omegaconf==2.3.0
opencv-python==4.5.5.62
packaging==23.2
Pillow==9.0.1
protobuf==4.24.4
psutil==5.9.7
pybind11==2.7.1
pytorch-lightning==1.9.5
PyYAML==6.0.1
requests==2.31.0
scipy==1.7.3
sentry-sdk==1.39.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
torch==1.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tqdm==4.62.3
typing_extensions==4.7.1
urllib3==2.0.7
wandb==0.16.1
yarl==1.9.4
zipp==3.15.0
aiohttp==3.8.6
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.3
asynctest==0.13.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cmake==3.21.0
commentjson==0.9.0
docker-pycreds==0.4.0
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.11
GitPython==3.1.40
Hydra==2.5
idna==3.6
imageio==2.16.1
importlib-metadata==6.7.0
lark-parser==0.7.8
lightning-utilities==0.10.0
multidict==6.0.4
networkx==2.6.3
numpy==1.21.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
omegaconf==2.3.0
opencv-python==4.5.5.62
packaging==23.2
Pillow==9.0.1
protobuf==4.24.4
psutil==5.9.7
pybind11==2.7.1
pytorch-lightning==1.9.5
PyWavelets==1.3.0
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.19.3
scipy==1.7.3
sentry-sdk==1.39.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
tifffile==2021.11.2
torch==1.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tqdm==4.62.3
typing_extensions==4.7.1
urllib3==2.0.7
wandb==0.16.1
yarl==1.9.4
zipp==3.15.0
